move filters to event_filters.py
